+++ components/Field/Field.stories.js
@@ -11,10 +11,12 @@ export default {
 
 export const Primary = () => {
   const [state, setState] = React.useState('');
+  const [text, setText] = React.useState('');
   return (
     <ThemeProvider theme={theme}>
       <div style={{margin: 20}}>
-        <Field placeholder="Digite alguma coisa" value={state} onChange={setState} validation={false} validationMessage="Teste"/>
+        <Field placeholder="Digite alguma coisa" enterKey={() => setText('click enter key')} value={state} onChange={setState} validation={false} validationMessage="Teste"/>
+        <pre>{text}</pre>
       </div>
     </ThemeProvider>
   );