+++ components/Field/index.js
@@ -1,6 +1,6 @@
 import {get} from 'lodash';
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, {useEffect} from 'react';
 import InputMask from 'react-input-mask';
 import {withTheme} from 'styled-components';
 
@@ -19,6 +19,7 @@ const Field = props => {
     styledMessage,
     styledInput,
     styledField,
+    styledRoot,
     onChange,
     onIconClick,
     type,
@@ -30,8 +31,18 @@ const Field = props => {
     on,
     id,
     name,
-    mask
+    mask,
+    enterKey
   } = props;
+
+  // Trigger to Handle enter keydown for forms
+  const handleKeyPress = event => {
+    if (event.keyCode === 13) enterKey(id);
+  };
+  useEffect(() => {
+    enterKey && window.addEventListener('keydown', handleKeyPress);
+    return () => enterKey && window.removeEventListener('keydown', handleKeyPress);
+  });
 
   const styledLabelDefaultProps = {
     fontSize: get(styledLabel, 'fontSize', '14px'),
@@ -56,10 +67,16 @@ const Field = props => {
   };
 
   const styledInputDefaultProps = {
-    $fontSize: get(styledInput, 'fontSize', '14px'),
-    $fontWeight: get(styledInput, 'fontWeight', 400),
-    $fontFamily: get(styledInput, 'fontFamily', 'secondary'),
     $color: get(styledInput, 'color', 'neutral5'),
+    $radius: get(styledInput, 'radius', 'default'),
+  };
+
+  const styledRootDefaultProps = {
+    $fontSize: get(styledRoot, 'fontSize', '14px'),
+    $fontWeight: get(styledRoot, 'fontWeight', 400),
+    $fontFamily: get(styledRoot, 'fontFamily', 'secondary'),
+    $color: get(styledRoot, 'color', 'neutral2'),
+    placeholderColor: get(styledRoot, 'placeholderColor'),
     $radius: get(styledInput, 'radius', 'default'),
   };
 
@@ -105,9 +122,8 @@ const Field = props => {
           mask={mask}
           placeholder={placeholder}
           validation={validation}
-          {...styledInputDefaultProps}
         >
-          <Input />
+          <Input {...styledRootDefaultProps} />
         </InputMask>
         {renderIcon()}
       </InputContainer>
@@ -152,11 +168,18 @@ Field.propTypes = {
    * Props de estilo para o elemento input HTML
    */
   styledInput: PropTypes.shape({
+    color: PropTypes.oneOf(colorProps),
+    radius: PropTypes.oneOf(['unset', 'default', 'alternative'])
+  }),
+  /**
+   * Props de estilo para o componente HTML input
+   */
+  styledRoot: PropTypes.shape({
     fontSize: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
     fontWeight: PropTypes.number,
     fontFamily: PropTypes.oneOf(['primary', 'secondary']),
-    color: PropTypes.oneOf(colorProps),
-    radius: PropTypes.oneOf(['unset', 'default', 'alternative'])
+    radius: PropTypes.oneOf(['unset', 'default', 'alternative']),
+    placeholderColor: PropTypes.string,
   }),
   /**
    * Props de estilo para o componente React Field
@@ -194,7 +217,7 @@ Field.propTypes = {
   /**
    * Permite assinalar um id para o elemento input raíz
    */
-  id: PropTypes.string,
+  id: PropTypes.string.isRequired,
   /**
    * Props que recebe o type do Field (prop nativa HTML)
    */
@@ -223,6 +246,10 @@ Field.propTypes = {
    * Permite a aplicação de uma máscara no input
    */
   mask: PropTypes.string,
+  /**
+   * Permite manipular ação de clique na tecla Enter/Return (recebe como parâmetro callback o id do Field)
+   */
+  enterKey: PropTypes.func,
 };
 
 export default withTheme(Field);